renaming existing interface rule name
