+++ lib/rules/prefer-interface-prefix.js
